Refactor NewPlanForm with mocks
