+++ src/mocks/mocksNewPlan.ts
@@ -1,0 +1,8 @@
+import { SelectorOptionsType } from 'src/types/Table';
+
+export const mockSelectorOptions: SelectorOptionsType[] = [
+    { name: 'Selecciona...', value: '' },
+    { name: 'Compra-Venta', value: 'buy-sell' },
+    { name: 'Proforma', value: 'proform' },
+    { name: 'Factura', value: 'invoice' },
+];
