+++ src/pages/plan/newPlan/components/NewPlanForm/components/NewPlanSubmenu/index.tsx
@@ -3,10 +3,10 @@ import { ArrowUpCircle } from 'src/assets/icons/ArrowUpCircle';
 
 type Props = {
     onSave: () => void;
-    saved: boolean;
+    isCreated: boolean;
 };
 
-export const NewPlanSubmenu: React.FC<Props> = ({ onSave, saved }) => {
+export const NewPlanSubmenu: React.FC<Props> = ({ onSave, isCreated }) => {
     return (
         <section className="flex items-center justify-between w-full px-4 py-3 border-b-2 bg-gray-100">
             <div className="flex items-center gap-1">
@@ -16,9 +16,9 @@ export const NewPlanSubmenu: React.FC<Props> = ({ onSave, saved }) => {
             <ul className="flex gap-1 text-terciary-color font-semibold text-sm">
                 <button
                     onClick={onSave}
-                    disabled={saved}
+                    disabled={isCreated}
                     className={`flex flex-col lg:flex-row justify-center items-center text-sm py-1 px-2 border border-gray-400 rounded shadow ${
-                        !saved ? 'hover:bg-gray-200' : ''
+                        !isCreated ? 'hover:bg-gray-200' : ''
                     }`}
                 >
                     <Save className=" w-5 h-5 pr-1" alt="arrow up circle" />
